removed regions class, added valid regions and regions map to region enum
